media-service/requirements.txt
fastapi==0.115.0
minio==7.2.8
python-multipart==0.0.9
python-dotenv==1.0.1
uvicorn==0.30.6